requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==9.5.0  # https://github.com/python-pillow/Pillow
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.4.0  # https://github.com/evansd/whitenoise
redis==4.5.5  # https://github.com/redis/redis-py

# Django
# ------------------------------------------------------------------------------
django==4.2.1  # pyup: < 4.2  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==2.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.7  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.2.0  # https://github.com/jazzband/django-redis

# OTP Support
# ------------------------------------------------------------------------------
django-two-factor-auth==1.15.2 # https://django-otp-official.readthedocs.io/en/stable/index.html
django-axes==6.0.1
django-ipware==5.0.0
django-user-sessions==2.0.0
geoip2==4.7.0
django-imagekit==4.1.0
django-richtextfield==1.6.1
django-simple-captcha==0.5.17
djangorestframework==3.14.0

django-storages==1.11.1
django-anymail==8.4
## The following requirements were added by pip freeze:
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alabaster==0.7.12
amqp==5.1.1
anyio==3.6.2
appnope==0.1.2
argon2-cffi==21.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.1
asgiref==3.6.0
astroid==2.8.3
asttokens==2.2.1
async-timeout==4.0.2
atomicwrites==1.4.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
binaryornot==0.4.4
black==21.9b0
boto3==1.18.51
botocore==1.21.51
celery==5.2.7
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
coverage==6.0.2
crispy-bootstrap5==0.7
cryptography==3.4.7
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.3
Django==4.2
django-allauth==0.51.0
django-appconf==1.0.5
django-bootstrap-form==3.4
django-celery-beat==2.4.0
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-coverage-plugin==2.0.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-environ==0.8.0
django-extensions==3.1.3
django-formtools==2.4
django-imagekit==4.0.2
django-ipware==4.0.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-ranged-response==0.2.0
django-redis==5.0.0
django-simple-captcha==0.5.14
django-smart-selects==1.5.9
django-stubs==1.13.1
django-stubs-ext==0.7.0
django-timezone-field==5.0
django-two-factor-auth==1.15.2
django-user-agents==0.4.0
django-user-sessions==2.0.0
djangorestframework==3.12.4
djangorestframework-stubs==1.7.0
docutils==0.17.1
drf-spectacular==0.24.2
environ==1.0
exceptiongroup==1.0.4
executing==1.2.0
factory-boy==3.2.0
Faker==9.5.2
filelock==3.3.0
flake8==4.0.1
flake8-isort==4.1.1
frozenlist==1.3.3
identify==2.3.1
idna==2.10
imagesize==1.2.0
inflection==0.5.1
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.28.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.2
jinja2-time==0.2.0
jmespath==0.10.0
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
multidict==6.0.4
mypy==0.910
mypy-extensions==0.4.3
nodeenv==1.6.0
numpy==1.21.3
oauthlib==3.1.1
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
phonenumberslite==8.13.11
pickleshare==0.7.5
pilkit==2.0
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
pre-commit==2.15.0
prompt-toolkit==3.0.21
protobuf==3.17.3
psycopg2==2.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.10.0
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
Pygments==2.10.0
PyJWT==2.1.0
pylint==2.11.1
pylint-celery==0.3
pylint-django==2.4.4
pylint-plugin-utils==0.6
pyparsing==3.0.1
pypng==0.20220715.0
pyrsistent==0.19.2
pytest==6.2.5
pytest-django==4.4.0
pytest-sugar==0.9.4
python-crontab==2.6.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
qrcode==7.4.2
redis==3.5.3
regex==2021.10.23
requests==2.25.1
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.1.0
Sphinx==4.2.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
stack-data==0.6.2
termcolor==1.1.0
testfixtures==6.18.3
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
tomlkit==0.11.6
tornado==6.1
traitlets==5.1.0
transifex-client==0.12.5
twilio==8.2.0
types-Markdown==3.4.2.1
types-pytz==2022.6.0.1
types-PyYAML==6.0.12.2
types-requests==2.28.11.5
types-urllib3==1.26.25.4
typing-extensions==3.10.0.2
tzdata==2022.7
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.5
user-agents==2.2.0
vine==5.0.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
watchdog==2.2.0
watchfiles==0.18.1
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.13.2
yarl==1.9.2
